[Alerting UI] Changed alerting UIs use new rule APIs. (#96018)
* [Alerting UI] Changed alerting UIs use new rule APIs.

* added unit tests

* fixed types

* fixed types

* fixed types

* fixed due to comments
