+++ x-pack/plugins/triggers_actions_ui/public/application/components/health_check.test.tsx
@@ -59,8 +59,13 @@ describe('health check', () => {
 
   it('renders children if keys are enabled', async () => {
     useKibanaMock().services.http.get = jest.fn().mockResolvedValue({
-      isSufficientlySecure: true,
-      hasPermanentEncryptionKey: true,
+      is_sufficiently_secure: true,
+      has_permanent_encryption_key: true,
+      alerting_framework_heath: {
+        decryption_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+        execution_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+        read_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+      },
       isAlertsAvailable: true,
     });
     const { queryByText } = render(
@@ -78,8 +83,13 @@ describe('health check', () => {
 
   test('renders warning if TLS is required', async () => {
     useKibanaMock().services.http.get = jest.fn().mockImplementation(async () => ({
-      isSufficientlySecure: false,
-      hasPermanentEncryptionKey: true,
+      is_sufficiently_secure: false,
+      has_permanent_encryption_key: true,
+      alerting_framework_heath: {
+        decryption_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+        execution_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+        read_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+      },
       isAlertsAvailable: true,
     }));
     const { queryAllByText } = render(
@@ -110,8 +120,13 @@ describe('health check', () => {
 
   test('renders warning if encryption key is ephemeral', async () => {
     useKibanaMock().services.http.get = jest.fn().mockImplementation(async () => ({
-      isSufficientlySecure: true,
-      hasPermanentEncryptionKey: false,
+      is_sufficiently_secure: true,
+      has_permanent_encryption_key: false,
+      alerting_framework_heath: {
+        decryption_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+        execution_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+        read_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+      },
       isAlertsAvailable: true,
     }));
     const { queryByText, queryByRole } = render(
@@ -139,8 +154,13 @@ describe('health check', () => {
 
   test('renders warning if encryption key is ephemeral and keys are disabled', async () => {
     useKibanaMock().services.http.get = jest.fn().mockImplementation(async () => ({
-      isSufficientlySecure: false,
-      hasPermanentEncryptionKey: false,
+      is_sufficiently_secure: false,
+      has_permanent_encryption_key: false,
+      alerting_framework_heath: {
+        decryption_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+        execution_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+        read_health: { status: 'ok', timestamp: '2021-04-01T21:29:22.991Z' },
+      },
       isAlertsAvailable: true,
     }));
